+++ analysiscoordinatorwiring.cpp
@@ -57,7 +57,6 @@ void AnalysisCoordinator::connectSchedulerSignals()
                 dependencies.handleWorkspaceSymbolAnalysisFinished(
                     filesAnalyzed,
                     totalSymbols);
-                refreshActiveEditorForFile(QString());
             });
     connect(scheduler,
             &AnalysisScheduler::relationshipAnalysisFinished,