+++ diagnosticsrefreshcontroller.cpp
@@ -1,21 +1,8 @@
 #include "diagnosticsrefreshcontroller.h"
 
-#include <QTimer>
-
 DiagnosticsRefreshController::DiagnosticsRefreshController(QObject* parent)
     : QObject(parent)
 {
-    refreshTimer = new QTimer(this);
-    refreshTimer->setSingleShot(true);
-    refreshTimer->setInterval(100);
-    connect(refreshTimer, &QTimer::timeout, this, [this]() {
-        const QString fileName = pendingFullRefresh
-            ? QString()
-            : pendingFileName;
-        pendingFileName.clear();
-        pendingFullRefresh = false;
-        emit diagnosticsRefreshRequested(fileName);
-    });
 }
 
 void DiagnosticsRefreshController::requestRefresh(const QString& fileName)
@@ -29,5 +16,19 @@ void DiagnosticsRefreshController::requestRefresh(const QString& fileName)
         pendingFullRefresh = true;
         pendingFileName.clear();
     }
-    refreshTimer->start();
+    if (refreshQueued)
+        return;
+    refreshQueued = true;
+    QMetaObject::invokeMethod(
+        this,
+        [this]() {
+            refreshQueued = false;
+            const QString fileName = pendingFullRefresh
+                ? QString()
+                : pendingFileName;
+            pendingFileName.clear();
+            pendingFullRefresh = false;
+            emit diagnosticsRefreshRequested(fileName);
+        },
+        Qt::QueuedConnection);
 }